added player inventory fields

+++ src/QtHelpers/TreeViewMemoryFields.cpp
@@ -13,7 +13,6 @@
 #include <iomanip>
 #include <sstream>
 
-
 S2Plugin::TreeViewMemoryFields::TreeViewMemoryFields(ViewToolbar* toolbar, MemoryMappedData* mmd, QWidget* parent) : QTreeView(parent), mToolbar(toolbar), mMemoryMappedData(mmd)
 {
     mHTMLDelegate = std::make_unique<HTMLDelegate>();
@@ -486,7 +485,7 @@ void S2Plugin::TreeViewMemoryFields::updateValueForField(const MemoryField& fiel
         {
             int16_t value = (memoryOffset == 0 ? 0 : Script::Memory::ReadWord(memoryOffset));
             itemValue->setData(QString::asprintf("%d", value), Qt::DisplayRole);
-            auto newHexValue = QString::asprintf("0x%04X", value);
+            auto newHexValue = QString::asprintf("0x%04X", static_cast<uint16_t>(value));
             itemField->setBackground(itemValueHex->data(Qt::DisplayRole) == newHexValue ? Qt::transparent : highlightColor);
             itemValueHex->setData(newHexValue, Qt::DisplayRole);
             itemValue->setData(value, gsRoleRawValue);
